liberty: overlay: Turn on button backlight for min lux value only
Soft button backlight is always controlled by light sensor, even
when auto-brightness is disabled. This change makes the backlight
turned on only with the minimum lux value of 33. When it goes to
the next lux value: 77, the environment should be bright enough
for us to see all the symbols/word of soft buttons without
backlight. This change potentially reduces battery consumption.

Change-Id: I0ffb490433571de3809bdb7998ca6277f2696dc5